pyproject.toml
[project]
name = "pdflab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "pyinstaller>=6.19.0",
    "pymupdf>=1.27.2.2",
]
